% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.fcs_to_rds.R
\name{fcs_to_rds}
\alias{fcs_to_rds}
\title{Converting FCS files to RDS files}
\usage{
fcs_to_rds(paths, batch_key = batch_key)
}
\arguments{
\item{paths}{Character vector of paths to store intput, intermediary results, outputs...}

\item{batch_key}{Data from plate-based experiments (batch_key="plate") or not (batch_key="batch")}
}
\value{
Generating fcs_metadata_df.rds and fcs_rawInten_mt.rds files.
}
\description{
This function has been designed to convert the raw FCS files to data matrix and export to RDS files.
}
\author{
Hsiao-Chi Liao
}
