% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegendscreenPipe.R
\docType{package}
\name{LegendscreenPipe}
\alias{LegendscreenPipe}
\alias{LegendscreenPipe-package}
\title{LegendscreenPipe: End-to-end toolboxs for analysing data from the LegendScreen assay, and fluorescence flow cytometry (FFC).}
\description{
This package provides functions for normalising single-cell proteomics data from LEGENDScreen assay as well as the fluorescence flow cytometry (FFC) assay, followed by cluster analysis.
}
\author{
Hsiao-Chi Liao \email{hsiaochi.liao@student.unimelb.edu.au}
}
