% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pipBackbone.R
\name{pipBackbone}
\alias{pipBackbone}
\title{The toolbox for analysing single-cell protein intensity from fluorescence flow cytometry (FFC).}
\usage{
pipBackbone(FCSpath="/PathToFCSfiles/",
Outpath="/PathToOutputFolder/Output/",
batch_key="plate",
bkb.v = c("FSC-H", "FSC-W", "SSC-H", "SSC-W", "CD69-CD301b", "MHCII", "CD4", "CD44", "CD8", "CD11c", "CD11b",
"F480", "Ly6C", "Lineage", "CD45a488", "CD24", "CD103"),
bkb.upper.quntile=0.9, bkb.lower.quntile=0.1, bkb.min.quntile=0.01,
trans.dat="cent.lgc")
}
\arguments{
\item{FCSpath}{Path to the input directory where input FCS files are stored (one file per well). Will look for FCS files recursively in that directory.}

\item{Outpath}{Path to the output directory where intermediate results and final results will be stored}

\item{batch_key}{Data from plate-based experiments (batch_key="plate") or not (batch_key="batch")}

\item{bkb.v}{A vector of the names of the backbone markers (MUST match to the names in the FCS file).}

\item{bkb.upper.quntile}{The cut-off (default = 0.9) for selecting cells used for estimating the parameter of signal.}

\item{bkb.lower.quntile}{The cut-off (default = 0.1) for selecting cells used for estimating the parameters of noise.}

\item{bkb.min.quntile}{The cut-off (default = 0.01) for omitting the cells with the smallest values to minimise the impact of outliers.}

\item{trans.dat}{Method for transforming raw intensities (default: Logicle).}
}
\value{
Normalised backbone measurements. Cluster analysis for normalised backbones. Graphs will be provided if set visualisation = TRUE.
}
\description{
This function is an end-to-end toolbox for analysing the data from FFC in FCS format, including data normalisation and cluster analysis.
}
\author{
Hsiao-Chi Liao
}
